requirements.txt
streamlit==1.18.0
google-generativeai==0.1.0
gtts==2.3.1
pillow==9.5.0
python-dotenv==1.0.0
